rota de listar todods os produtos

+++ src/controllers/ProductController.js
@@ -6,6 +6,23 @@ import fs from 'fs';
 import QrcodeProduct from "../models/QrcodeProduct";
 
 class ProductController {
+  async index(req, res) {
+    const { admin } = req.user;
+
+    if (!admin) {
+      return res.status(401).json({ error: "validations fails" });
+    }
+
+    try {
+      await Product.find({ sold: false })
+      .populate('image')
+      .exec(function(_, usersDocuments) {
+        return res.status(200).json(usersDocuments);
+      });
+    }catch(err) {
+      return res.status(400).json(err);
+    }
+}
   async store(req, res) {
     /** PARA TESTAR NO INSOMIA: SALVAR PRIMEIRO UMA IMAGEM E USAR ID DA IMAGEM EM IMAGE */
 
@@ -16,6 +33,12 @@ class ProductController {
       quantity: Yup.number().required().positive().integer(),
       image: Yup.string().required(),
     });
+
+    const { admin } = req.user;
+    
+    if (!admin) {
+      return res.status(401).json({ error: "validations fails" });
+    }
 
     const checkSchema = await schemaValidation.isValid(req.body);
 
@@ -60,6 +83,12 @@ class ProductController {
       quantity: Yup.number().positive().integer(),
     });
 
+    const { admin } = req.user;
+    
+    if (!admin) {
+      return res.status(401).json({ error: "validations fails" });
+    }
+
     const { id } = req.params;
 
     const checkSchema = await schemaValidation.isValid(req.body);
@@ -81,6 +110,12 @@ class ProductController {
   async delete(req, res) {
     try {
       const { id } = req.params;
+
+      const { admin } = req.user;
+    
+      if (!admin) {
+        return res.status(401).json({ error: "validations fails" });
+      }
 
       const { image } = await Product.findById({ _id: id });
       await ImageProduct.findByIdAndDelete(image);
